RepositoryImplの廃止
