+++ src/main/java/com/ngnmsn/template/domain/service/LoginUserDetailService.java
@@ -2,7 +2,7 @@ package com.ngnmsn.template.domain.service;
 
 import com.ngnmsn.template.domain.model.auth.AuthResult;
 import com.ngnmsn.template.domain.model.auth.LoginUserDetails;
-import com.ngnmsn.template.repository.impl.AuthRepositoryImpl;
+import com.ngnmsn.template.repository.AuthRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
@@ -16,11 +16,11 @@ import org.springframework.stereotype.Service;
 public class LoginUserDetailService implements UserDetailsService {
 
   @Autowired
-  private AuthRepositoryImpl authRepositoryImpl;
+  private AuthRepository authRepository;
 
   @Override
   public UserDetails loadUserByUsername(String loginId) throws UsernameNotFoundException {
-    AuthResult result = authRepositoryImpl.findByLoginId(loginId);
+    AuthResult result = authRepository.findByLoginId(loginId);
     return new LoginUserDetails(result);
   }
 }
